Unity2017にアップデートしたのでコミ

+++ Assets/Scripts/MetricCalculation/MeterUITxt.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public abstract class MeterUITxt : MonoBehaviour {
@@ -8,7 +7,7 @@ public abstract class MeterUITxt : MonoBehaviour {
     protected UnityEngine.UI.Text TextUI;
     [SerializeField]
     protected float UpdateSecond = 1.0f;
-    protected float Metrics;
+    protected double Metrics;
     protected IEnumerator Coroutine;
 
     // Use this for initialization
